Added code to file 102-fibonacci.c v4

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,17 +9,17 @@
 
 int main(void)
 {
-int fib1 = 1, fib2 = 2, fib3 = 3, num;
+long int fib1 = 1, fib2 = 2, fib3 = 3, num;
 
-printf("%d, ", fib1);
-printf("%d, ", fib2);
+printf("%ld, ", fib1);
+printf("%ld, ", fib2);
 for (num = 3; num < 50; num++)
 {
-printf("%d, ", fib3);
+printf("%ld, ", fib3);
 fib1 = fib2;
 fib2 = fib3;
 fib3 = fib1 + fib2;
 }
-printf("%d\n", fib3);
+printf("%ld\n", fib3);
 return (0);
 }
